chatBot/requirements.txt
flask
flask-cors
openpyxl
pandas
python-dotenv
requests
beautifulsoup4
thefuzz
fuzzywuzzy
langchain
chromadb
openai
langchain-openai
langchain-community